move arm op kernels to central_arm_func
